refactor(firestore): extract transaction execution
